refactor: moved data dir functions to own file
